requirements.txt
scikit-learn
tqdm==4.64.1
transformers==4.26.1
datasets==2.10.0
evaluate==0.4.0
torch
matplotlib
sentencepiece

faiss-cpu
python-dotenv
openai
